Add files via upload
In this zip file you will find the html and css code... with the images of the dice which has been inserted on the website. For doing practise of javascript you can download these zip files and first try coding js with yourself to match mine website and if unable to do then can take reference from my uploaded one.